+++ src/TimePicker/__stories__/TimePicker.stories.js
@@ -1,4 +1,5 @@
 /* eslint-disable react/no-multi-comp */
+import FormLabel from '../../Forms/FormLabel';
 import React from 'react';
 import TimePicker from '../TimePicker';
 
@@ -7,21 +8,61 @@ export default {
     component: TimePicker
 };
 export const primary = () => (
-    <TimePicker />
+    <>
+        <FormLabel htmlFor='timepickerPrimaryExample-input' >Time</FormLabel>
+        <TimePicker
+            buttonProps={{
+                'aria-label': 'Show/hide time picker'
+            }}
+            id='timepickerPrimaryExample' />
+    </>
 );
 
 export const disabled = () => (
-    <TimePicker disabled />
+    <>
+        <FormLabel htmlFor='timepickerDisabledExample-input' >Time</FormLabel>
+        <TimePicker
+            buttonProps={{
+                'aria-label': 'Show/hide time picker'
+            }}
+            disabled
+            id='timepickerDisabledExample' />
+    </>
 );
 
 export const format12Hours = () => (
-    <TimePicker format12Hours />
+    <>
+        <FormLabel htmlFor='timepicker12HrExample-input' >Time</FormLabel>
+        <TimePicker
+            buttonProps={{
+                'aria-label': 'Show/hide time picker'
+            }}
+            format12Hours
+            id='timepicker12HrExample' />
+    </>
 );
 
 export const initialValue = () => (
-    <TimePicker format12Hours value='10:30:34 am' />
+    <>
+        <FormLabel htmlFor='timepickerInitExample-input' >Time</FormLabel>
+        <TimePicker
+            buttonProps={{
+                'aria-label': 'Show/hide time picker'
+            }}
+            format12Hours
+            id='timepickerInitExample'
+            value='10:30:34 am' />
+    </>
 );
 
 export const hideSeconds = () => (
-    <TimePicker showSecond={false} />
+    <>
+        <FormLabel htmlFor='timepickerHideSecondExample-input' >Time</FormLabel>
+        <TimePicker
+            buttonProps={{
+                'aria-label': 'Show/hide time picker'
+            }}
+            id='timepickerHideSecondExample'
+            showSecond={false} />
+    </>
 );